clean up and implement reversed view
