+++ core/src/main/scala/com/raphtory/internals/graph/GraphLens.scala
@@ -1,3 +1,3 @@
 package com.raphtory.internals.graph
 
-abstract private[raphtory] class GraphLens(jobId: String, start: Long, end: Long)
+abstract private[raphtory] class GraphLens(jobId: String, start: Long, end: Long) extends LensInterface
