+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/PojoExplodedEdge.scala
@@ -1,10 +1,9 @@
 package com.raphtory.internals.storage.pojograph.entities.external
 
-import com.raphtory.api.analysis.visitor.ConcreteExplodedEdge
 import com.raphtory.internals.storage.pojograph.PojoGraphLens
 import com.raphtory.internals.storage.pojograph.entities.internal.PojoEdge
 
-private[raphtory] class PojoExplodedEdge(
+private[pojograph] class PojoExplodedEdge(
     val edge: PojoEdge,
     val view: PojoGraphLens,
     override val ID: Long,
@@ -32,7 +31,7 @@ private[raphtory] class PojoExplodedEdge(
       new PojoExplodedInOutEdge(other, this, asInEdge)
 }
 
-private[raphtory] object PojoExplodedEdge {
+private[pojograph] object PojoExplodedEdge {
 
   def fromEdge(pojoExEdge: PojoExEdge, timestamp: Long): PojoExplodedEdge =
     new PojoExplodedEdge(
@@ -40,6 +39,45 @@ private[raphtory] object PojoExplodedEdge {
             pojoExEdge.view,
             pojoExEdge.ID,
             pojoExEdge.start,
+            timestamp
+    )
+}
+
+private[pojograph] class PojoExplodedInOutEdge(
+    in: PojoExplodedEdge,
+    out: PojoExplodedEdge,
+    asInEdge: Boolean = false
+) extends PojoExInOutEdgeBase[PojoExplodedInOutEdge, PojoExplodedEdge, Long](in, out, asInEdge)
+        with PojoExplodedEdgeBase[Long] {
+
+  override def timestamp: Long = in.timestamp
+}
+
+private[pojograph] class PojoReversedExplodedEdge(
+    objectEdge: PojoEdge,
+    override val view: PojoGraphLens,
+    override val ID: Long,
+    override val start: Long,
+    override val timestamp: Long
+) extends PojoExplodedEdge(objectEdge, view, ID, start, timestamp) {}
+
+private[pojograph] object PojoReversedExplodedEdge {
+
+  def fromExplodedEdge(pojoExEdge: PojoExplodedEdge, timestamp: Long): PojoReversedExplodedEdge =
+    new PojoReversedExplodedEdge(
+            pojoExEdge.edge,
+            pojoExEdge.view,
+            pojoExEdge.ID,
+            pojoExEdge.start,
+            timestamp
+    )
+
+  def fromReversedEdge(pojoExReversedEdge: PojoExReversedEdge, timestamp: Long): PojoReversedExplodedEdge =
+    new PojoReversedExplodedEdge(
+            pojoExReversedEdge.edge,
+            pojoExReversedEdge.view,
+            pojoExReversedEdge.ID,
+            pojoExReversedEdge.start,
             timestamp
     )
 }
