+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/PojoVertexBase.scala
@@ -60,6 +60,8 @@ private[raphtory] trait PojoConcreteVertexBase[T] extends PojoVertexBase {
 
   def viewUndirected: PojoUndirectedVertexView[T]
 
+  def viewReversed: PojoReversedVertexView[T]
+
   // messaging
   def hasMessage: Boolean =
     multiQueue.getMessageQueue(lens.superStep).nonEmpty
